fix: deleted lib.rs on accident
